Do not start OpenVPNService at install time
Startup mode had been set to "manual" earlier, but "openvpnserv2.exe -install"
option still attempted to launch the service. This failed if tap-windows6 was
not yet usable, as is the case when using the official OpenVPN installers. This
commit also bumps version to 1.1.0.0 to make it clear that the behavior has
changed.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>